% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrstdlcrt.R
\name{print.mrs}
\alias{print.mrs}
\title{Print method for mrs objects}
\usage{
\method{print}{mrs}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"mrs"}.}

\item{...}{Unused.}
}
\value{
\code{x} invisibly.
}
\description{
Print method for mrs objects
}
