% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrstdlcrt.R
\name{plot.mrs}
\alias{plot.mrs}
\title{Plot method for mrs objects}
\usage{
\method{plot}{mrs}(x, level = 0.95, estimand = NULL, point_size = 2.8, ...)
}
\arguments{
\item{x}{An object of class \code{"mrs"}.}

\item{level}{Confidence level.}

\item{estimand}{Subset of estimands to plot.}

\item{point_size}{Point size.}

\item{...}{Unused.}
}
\value{
ggplot object invisibly.
}
\description{
Plot method for mrs objects
}
